Reorganize drivers for workflow components
